restructure LIst stories
