+++ vuu-ui/packages/showcase/src/examples/UIControls/Combobox.stories.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { ComboBox } from '@vuu-ui/ui-controls';
-import { usa_states } from './List.data';
+import { usa_states } from './List/List.data';
 
 export default {
   title: 'UI Controls/ComboBox',
