+++ vuu-ui/packages/showcase/src/examples/UIControls/List/uncontrolled/List.uncontrolled.data-driven.stories.jsx
@@ -8,7 +8,8 @@ import {
   ListItemHeader,
   VirtualizedList
 } from '@vuu-ui/ui-controls';
-import { usa_states, usa_states_cities, groupByInitialLetter, random_1000 } from './List.data';
+
+import { usa_states, usa_states_cities, groupByInitialLetter, random_1000 } from '../List.data';
 
 import '@vuu-ui/theme/index.css';
 
@@ -404,93 +405,6 @@ export const ExtendedSelectList = () => {
   );
 };
 
-export const ControlledList = () => {
-  const [selected, setSelected] = useState([]);
-  const [hilitedIdx, setHilitedIdx] = useState(-1);
-
-  const handleChangeController = (evt, newSelected) => {
-    console.log(`handleChangeController`);
-    setSelected(newSelected);
-  };
-  const handleChangeControlled = (idx) => {
-    console.log(`handleChangeControlled`);
-    console.log(`controlled clicked ${idx}`);
-  };
-
-  return (
-    <div style={{ display: 'flex', height: 600 }}>
-      <div>
-        <input type="text" />
-        <List
-          id="controller"
-          source={usa_states}
-          onChange={handleChangeController}
-          onHighlight={(idx) => setHilitedIdx(idx)}
-        />
-        <input type="text" />
-      </div>
-      <div>
-        <input type="text" />
-        <List
-          id="controlled"
-          highlightedIdx={hilitedIdx}
-          selected={selected}
-          source={usa_states}
-          onChange={handleChangeControlled}
-        />
-        <input type="text" />
-      </div>
-    </div>
-  );
-};
-
-export const FullyControlledList = () => {
-  const [selected, setSelected] = useState([]);
-  const [hilitedIdx, setHilitedIdx] = useState(-1);
-
-  const handleChangeController = (evt, newSelected) => {
-    console.log(`handleChangeController`);
-    setSelected(newSelected);
-  };
-  const handleChangeControlled = (idx) => {
-    console.log(`handleChangeControlled`);
-    console.log(`controlled clicked ${idx}`);
-  };
-
-  const moveUp = () => {
-    setHilitedIdx((val) => Math.max(0, val - 1));
-  };
-
-  const selectCurrent = () => {
-    const [selectedIdx] = selected;
-    const newSelection = selectedIdx === hilitedIdx || hilitedIdx === -1 ? [] : [hilitedIdx];
-    setSelected(newSelection);
-  };
-
-  const moveDown = () => {
-    setHilitedIdx((val) => Math.min(usa_states.length - 1, val + 1));
-  };
-
-  return (
-    <div style={{ height: 600 }}>
-      <div style={{ display: 'flex', gap: 12 }}>
-        <Button onClick={moveDown}>Highlight down</Button>
-        <Button onClick={moveUp}>Highlight up</Button>
-        <Button onClick={selectCurrent}>Select</Button>
-      </div>
-      <div style={{ height: 600 }}>
-        <List
-          id="controlled"
-          highlightedIdx={hilitedIdx}
-          selected={selected}
-          source={usa_states}
-          onChange={handleChangeControlled}
-        />
-      </div>
-    </div>
-  );
-};
-
 export const VirtualizedExample = () => {
   const data = useMemo(() => {
     const data = [];
@@ -512,255 +426,6 @@ export const VirtualizedExample = () => {
       <VirtualizedList source={data} />
       <input type="text" />
     </div>
-  );
-};
-
-export const DeclarativeList = () => {
-  const [selectedValue, setSelectedValue] = useState('');
-  console.log(`render example`);
-  return (
-    <>
-      <input type="text" />
-      <div
-        style={{
-          fontFamily: 'Roboto',
-          width: 150,
-          height: 400,
-          position: 'relative',
-          border: 'solid 1px #ccc'
-        }}>
-        <List onChange={(value) => setSelectedValue(value)}>
-          <ListItem>Value 1</ListItem>
-          <ListItem>Value 2</ListItem>
-          <ListItem>Value 3</ListItem>
-          <ListItem>Value 4</ListItem>
-        </List>
-      </div>
-      <input type="text" />
-      <div>{usa_states[selectedValue]}</div>
-    </>
-  );
-};
-
-export const DeclarativeListUsingDivs = () => {
-  const [selectedValue, setSelectedValue] = useState('');
-  console.log(`render example`);
-  return (
-    <>
-      <input type="text" />
-      <div
-        style={{
-          fontFamily: 'Roboto',
-          width: 150,
-          height: 400,
-          position: 'relative',
-          border: 'solid 1px #ccc'
-        }}>
-        <List onChange={(value) => setSelectedValue(value)}>
-          <div>
-            <span>Value 1</span>
-          </div>
-          <div>
-            <span>Value 2</span>
-          </div>
-          <div>Value 3</div>
-          <div>Value 4</div>
-        </List>
-      </div>
-      <input type="text" />
-      <div>{usa_states[selectedValue]}</div>
-    </>
-  );
-};
-
-export const DeclarativeListWithHeadersUsingDivs = () => {
-  const [selectedValue, setSelectedValue] = useState('');
-  console.log(`render example`);
-  return (
-    <>
-      <input type="text" />
-      <div
-        style={{
-          fontFamily: 'Roboto',
-          width: 150,
-          height: 400,
-          position: 'relative',
-          border: 'solid 1px #ccc'
-        }}>
-        <List onChange={(value) => setSelectedValue(value)} collapsibleHeaders>
-          <div data-header label="Group 1" />
-          <div>
-            <span>Value 1</span>
-          </div>
-          <div>
-            <span>Value 2</span>
-          </div>
-          <div data-header>Group 2</div>
-          <div>Value 3</div>
-          <div>Value 4</div>
-        </List>
-      </div>
-      <input type="text" />
-      <div>{usa_states[selectedValue]}</div>
-    </>
-  );
-};
-
-export const DeclarativeListWithGroups = () => {
-  const [selectedValue, setSelectedValue] = useState('');
-  console.log(`render example`);
-  return (
-    <>
-      <input type="text" />
-      <div
-        style={{
-          fontFamily: 'Roboto',
-          width: 150,
-          height: 400,
-          maxHeight: 400,
-          position: 'relative',
-          border: 'solid 1px #ccc'
-        }}>
-        <List onChange={(value) => setSelectedValue(value)} collapsibleHeaders stickyHeaders>
-          <ListItemGroup title="Group 1">
-            <ListItem>Value 1.1</ListItem>
-            <ListItem>Value 1.2</ListItem>
-            <ListItem>Value 1.3</ListItem>
-            <ListItem>Value 1.4</ListItem>
-          </ListItemGroup>
-          <ListItemGroup title="Group 2">
-            <ListItem>Value 2.1</ListItem>
-            <ListItem>Value 2.2</ListItem>
-            <ListItem>Value 2.3</ListItem>
-            <ListItem>Value 2.4</ListItem>
-          </ListItemGroup>
-          <ListItemGroup title="Group 3">
-            <ListItem>Value 3.1</ListItem>
-            <ListItem>Value 3.2</ListItem>
-            <ListItem>Value 3.3</ListItem>
-            <ListItem>Value 3.4</ListItem>
-          </ListItemGroup>
-          <ListItemGroup title="Group 4">
-            <ListItem>Value 4.1</ListItem>
-            <ListItem>Value 4.2</ListItem>
-            <ListItem>Value 4.3</ListItem>
-            <ListItem>Value 4.4</ListItem>
-            <ListItem>Value 4.5</ListItem>
-            <ListItem>Value 4.6</ListItem>
-            <ListItem>Value 4.7</ListItem>
-          </ListItemGroup>
-          <ListItemGroup title="Group 5">
-            <ListItem>Value 5.1</ListItem>
-            <ListItem>Value 5.2</ListItem>
-            <ListItem>Value 5.3</ListItem>
-            <ListItem>Value 5.4</ListItem>
-          </ListItemGroup>
-        </List>
-      </div>
-      <input type="text" />
-      <div>{usa_states[selectedValue]}</div>
-    </>
-  );
-};
-
-export const DeclarativeListWithNestedGroups = () => {
-  const [selectedValue, setSelectedValue] = useState('');
-  console.log(`render example`);
-  return (
-    <>
-      <input type="text" />
-      <div
-        style={{
-          fontFamily: 'Roboto',
-          width: 150,
-          height: 400,
-          maxHeight: 400,
-          position: 'relative',
-          border: 'solid 1px #ccc'
-        }}>
-        <List onChange={(value) => setSelectedValue(value)} collapsibleHeaders stickyHeaders>
-          <ListItemGroup title="Group 1">
-            <ListItem>Value 1.1</ListItem>
-            <ListItem>Value 1.2</ListItem>
-            <ListItem>Value 1.3</ListItem>
-            <ListItem>Value 1.4</ListItem>
-          </ListItemGroup>
-          <ListItemGroup title="Group 2">
-            <ListItemGroup title="Group 2.1">
-              <ListItem>Value 2.1.1</ListItem>
-              <ListItem>Value 2.1.2</ListItem>
-              <ListItem>Value 2.1.3</ListItem>
-              <ListItem>Value 2.1.4</ListItem>
-            </ListItemGroup>
-            <ListItem>Value 2.2</ListItem>
-            <ListItem>Value 2.3</ListItem>
-            <ListItem>Value 2.4</ListItem>
-          </ListItemGroup>
-          <ListItemGroup title="Group 3">
-            <ListItem>Value 3.1</ListItem>
-            <ListItem>Value 3.2</ListItem>
-            <ListItem>Value 3.3</ListItem>
-            <ListItem>Value 3.4</ListItem>
-          </ListItemGroup>
-          <ListItemGroup title="Group 4">
-            <ListItem>Value 4.1</ListItem>
-            <ListItem>Value 4.2</ListItem>
-            <ListItem>Value 4.3</ListItem>
-            <ListItem>Value 4.4</ListItem>
-            <ListItem>Value 4.5</ListItem>
-            <ListItem>Value 4.6</ListItem>
-            <ListItem>Value 4.7</ListItem>
-          </ListItemGroup>
-          <ListItemGroup title="Group 5">
-            <ListItem>Value 5.1</ListItem>
-            <ListItem>Value 5.2</ListItem>
-            <ListItem>Value 5.3</ListItem>
-            <ListItem>Value 5.4</ListItem>
-          </ListItemGroup>
-        </List>
-      </div>
-      <input type="text" />
-      <div>{usa_states[selectedValue]}</div>
-    </>
-  );
-};
-
-export const DeclarativeListWithHeaders = () => {
-  const [selectedValue, setSelectedValue] = useState('');
-  console.log(`render example`);
-  return (
-    <>
-      <input type="text" />
-      <div
-        style={{
-          fontFamily: 'Roboto',
-          width: 150,
-          height: 400,
-          maxHeight: 400,
-          position: 'relative',
-          border: 'solid 1px #ccc'
-        }}>
-        <List stickyHeaders collapsibleHeaders onChange={(value) => setSelectedValue(value)}>
-          <ListItemHeader id="1">Group 1</ListItemHeader>
-          <ListItem>Value 1.1</ListItem>
-          <ListItem>Value 1.2</ListItem>
-          <ListItem>Value 1.3</ListItem>
-          <ListItem>Value 1.4</ListItem>
-          <ListItemHeader id="2">Group 2</ListItemHeader>
-          <ListItem>Value 2.1</ListItem>
-          <ListItem>Value 2.2</ListItem>
-          <ListItem>Value 2.3</ListItem>
-          <ListItem>Value 2.4</ListItem>
-          <ListItemHeader id="3">Group 3</ListItemHeader>
-          <ListItem>Value 3.1</ListItem>
-          <ListItem>Value 3.2</ListItem>
-          <ListItem>Value 3.3</ListItem>
-          <ListItem>Value 3.4</ListItem>
-        </List>
-      </div>
-      <input type="text" />
-      <div>{usa_states[selectedValue]}</div>
-    </>
   );
 };
