+++ vuu-ui/packages/showcase/src/examples/UIControls/Tree.stories.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { ComponentAnatomy } from '@heswell/component-anatomy';
 
 import { Tree, useItemsWithIds } from '@vuu-ui/ui-controls';
-import { usa_states_cities, groupByInitialLetter } from './List.data';
+import { usa_states_cities, groupByInitialLetter } from './List/List.data';
 import { folderData } from './Tree.data';
 
 export default {
