+++ vuu-ui/packages/showcase/src/examples/UIControls/index.js
@@ -6,5 +6,5 @@ export * as DropdownList from './DropdownList.stories';
 export * as Form from './Form.stories';
 export * as Menu from './Menu.stories';
 export * as Pill from './Pill.stories';
-export * as List from './List.stories';
+export * as List from './List';
 export * as Tree from './Tree.stories';
